requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
PyJWT
passlib
sqlmodel
python-dotenv
pydantic-settings